+++ gauntlet/packages/gauntlet-solana-contracts/src/commands/contracts/accessController/initialize.ts
@@ -1,5 +1,5 @@
 import { Result } from '@chainlink/gauntlet-core'
-import { SolanaCommand, TransactionResponse } from '@chainlink/gauntlet-solana'
+import { SolanaCommand, TransactionResponse } from '@chainlink/solana-gauntlet'
 import { Keypair, SystemProgram, SYSVAR_RENT_PUBKEY } from '@solana/web3.js'
 import { CONTRACT_LIST, getContract } from '../../../lib/contracts'