+++ gauntlet/packages/gauntlet-solana-contracts/src/commands/contracts/ocr2/transmit.dev.ts
@@ -1,5 +1,5 @@
 import { Result } from '@chainlink/gauntlet-core'
-import { SolanaCommand, TransactionResponse } from '@chainlink/gauntlet-solana'
+import { SolanaCommand, TransactionResponse } from '@chainlink/solana-gauntlet'
 import { parseIdlErrors, ProgramError, utils } from '@project-serum/anchor'
 import { Keypair, PublicKey, Transaction, TransactionInstruction } from '@solana/web3.js'
 import { createHash } from 'crypto'