+++ relayer/pkg/monitoring/chain_reader.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/gagliardetto/solana-go"
 	"github.com/gagliardetto/solana-go/rpc"
-	pkgSolana "github.com/smartcontractkit/chainlink-solana/pkg/solana"
+	pkgSolana "github.com/smartcontractkit/chainlink-solana/relayer/pkg/solana"
 )
 
 type ChainReader interface {