+++ relayer/pkg/monitoring/source_envelope.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/gagliardetto/solana-go/rpc"
 	relayMonitoring "github.com/smartcontractkit/chainlink-relay/pkg/monitoring"
-	pkgSolana "github.com/smartcontractkit/chainlink-solana/pkg/solana"
+	pkgSolana "github.com/smartcontractkit/chainlink-solana/relayer/pkg/solana"
 	"github.com/smartcontractkit/libocr/offchainreporting2/types"
 	"go.uber.org/multierr"
 )